requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
certifi==2020.6.20
chardet==3.0.4
decorator==4.4.2
Django==3.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
django-filter==2.3.0
django-ipware==3.0.1
django-phonenumber-field==5.0.0
django-summernote==0.8.11.6
djangorestframework==3.11.1
drf-recaptcha==2.0.3
idna==2.10
ipython==7.17.0
ipython-genutils==0.2.0
jedi==0.17.2
Markdown==3.2.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prompt-toolkit==3.0.6
ptyprocess==0.6.0
pyactiveresource==2.2.1
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ShopifyAPI==8.0.2
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.2.5
